SpikeTrainUtilities

Documentation for SpikeTrainUtilities.

SpikeTrainUtilities.SpikeTrains

SpikeTrainUtilities.bin_spikes

SpikeTrainUtilities.bin_spikes

SpikeTrainUtilities.bin_spikes

SpikeTrainUtilities.bin_spikes

SpikeTrainUtilities.circular_shift_of_spiketrains

SpikeTrainUtilities.circular_shift_of_spiketrains

SpikeTrainUtilities.draw_spike_raster

SpikeTrainUtilities.filter_out_small_isi!

SpikeTrainUtilities.filter_out_small_isi!

SpikeTrainUtilities.get_instantaneous_firing_rates

SpikeTrainUtilities.get_line_segments

SpikeTrainUtilities.line_segments_to_centers_and_size

SpikeTrainUtilities.line_segments_to_xynans

SpikeTrainUtilities.numerical_rates

SpikeTrainUtilities.sort_units!

SpikeTrainUtilities.SpikeTrains — Type

SpikeTrains(nunits::N,trains::Vector{Vector{R}},tstart::R,t_end::R) where {R,N}

Constructs a SpikeTrains object.

SpikeTrainUtilities.bin_spikes — Method

bin_spikes(S::SpikeTrains{R,N},dt::R) where {R,N}

Arguments

S : SpikeTrains object

dt : time bin size

Returns

timesbin::Vector{R} : `timesbin[k]is the midpoint of the timebink(i.e.S.t_start + (k-1/2)*dt`)

binned_spikes::Matrix{<:Integer} : binned_spikes[k,i] is the number of spikes that occur in the timebin k for neuron i (i.e. between S.t_start + (k-1)*dt and S.t_start + k*dt)

SpikeTrainUtilities.bin_spikes — Method

bin_spikes(S::SpikeTrains{R,N},binvector::Vector{R}) where {R,N}

Arguments

S : SpikeTrains object

binvector : vector of bin edges

Returns

timesbin::Vector{R} : `timesbin[k]is the midpoint of the timebink(i.e.binvector[k] + (binvector[k+1]-binvector[k])/2`)

binned_spikes::Matrix{<:Integer} : binned_spikes[k,i] is the number of spikes that occur in the timebin k for neuron i (i.e. between binvector[k] and binvector[k+1])

SpikeTrainUtilities.bin_spikes — Method

bin_spikes(Y::Vector{R},binvector::Vector{R}) where R

Arguments

Y::Vector{<:Real} : vector of spike times

binvector::AbstractVector{<:Real} : vector of bin edges

Returns

times_bin::Vector{R} : times_bin[k] is the midpoint of the timebin k (i.e. binvector[k] + (binvector[k+1]-binvector[k])/2)

binned_spikes::Vector{<:Integer} : binned_spikes[k] is the number of spikes that occur in the timebin k (i.e. between binvector[k] and binvector[k+1])

SpikeTrainUtilities.bin_spikes — Method

bin_spikes(Y::Vector{R},dt::R,Tend::R;Tstart::R=0.0) where R

Arguments

Y::Vector{<:Real} : vector of spike times

dt::Real : time bin size

Tend::Real : end time for the raster

Optional argument

Tstart::Real=0.0 : start time for the raster

Returns

timesbin::Vector{R} : `timesbin[k]is the midpoint of the timebink(i.e.Tstart + (k-1/2)*dt`)

binned_spikes::Vector{<:Integer} : binned_spikes[k] is the number of spikes that occur in the timebin k (i.e. between Tstart + (k-1)*dt and Tstart + k*dt)

SpikeTrainUtilities.circular_shift_of_spiketrains — Method

function circularshiftofspiketrains(sq::SpikeQuantity; shiftquantity::Bool=true)

Perform a circular shuffle of the event times in sq. This function randomly shifts each event time series by a random amount, wrapping around the duration of the SpikeQuantity object. If shift_quantity is true (default), the corresponding ys values are shifted along with their event times, maintaining their association after sorting. If shift_quantity is false, event_times are shifted and sorted, but the ys vectors for each unit are copied from the original SpikeQuantity without reordering. Thus, the i-th value in new_ys[k] will correspond to the i-th time in the newly sorted new_event_times[k], which may differ from its original association if the circular shift altered the temporal order of events.

The function returns a new SpikeQuantity object with the shuffled event times (and potentially ys), and the random shifts applied to each series. The random shifts are generated uniformly in the range [0, duration), where duration is the total duration of the SpikeQuantity.

Arguments

sq::SpikeQuantity: The input spike quantity to be shuffled. It must have a positive duration.

shift_quantity::Bool: If true (default), ys values follow their corresponding event times through the shift and sort. If false, ys vectors are copied directly from the input sq for each unit without reordering relative to the new sorted event times.

Returns

New SpikeQuantity object with the shuffled event times and ys.

Vector of random shifts applied to each event time series.

SpikeTrainUtilities.circular_shift_of_spiketrains — Method

function circularshiftof_spiketrains(spk::SpikeTrains)

Perform a circular shuffle of the spike trains in spk. This function randomly shifts each spike train by a random amount, wrapping around the duration of the spike trains. The function returns a new SpikeTrains object with the shuffled spike times, and the random shifts applied to each train. The random shifts are generated uniformly in the range [0, duration), where duration is the total duration of the spike trains.

Arguments

spk::SpikeTrains: The input spike trains to be shuffled. It must have a positive duration.

Returns

New SpikeTrains object with the shuffled spike times.

Vector of random shifts applied to each spike train.

SpikeTrainUtilities.draw_spike_raster — Method

drawspikeraster(trains::Vector{Vector{Float64}}, dt::Real,Tend::Real; Tstart::Real=0.0, spikesize::Integer = 5, spikeseparator::Integer = 1, backgroundcolor::Color=RGB(1.,1.,1.), spikecolors::Union{C,Vector{C}}=RGB(0.,0.0,0.0), max_size::Real=1E4) where C<:Color

Draws a matrix that contains the raster plot of the spike train.

Arguments

Trains : Vector of spike trains. The order of the vector corresponds to the order of the plot. First element is at the top, second is second row, etc.

dt : time interval representing one horizontal pixel

Tend : final time to be considered

Optional arguments

Tstart::Real : starting time

spike_size::Integer : heigh of spike (in pixels)

spike_separator::Integer : space between spikes, and vertical padding

background_color::Color : self-explanatory

spike_colors::Union{Color,Vector{Color}} : if a single color, color of all spikes, if vector of colors, color for each neuron (length should be same as number of neurons)

max_size::Integer : throws an error if image is larger than this number (in pixels)

Returns

raster_matrix::Matrix{Color} you can save it as a png file

SpikeTrainUtilities.filter_out_small_isi! — Method

function filteroutsmallisi!(spiketrains::SpikeTrains, dtmin::Real)

Filter out spikes (events) that are closer than dt_min from each other in the spike trains object This modifies the input spiketrains in place, removing spikes that are too close together.

Arguments:

spiketrains::SpikeTrains: An instance of SpikeTrains, which contains multiple spike trains.

dt_min::Real: The minimum time difference between spikes. Spikes closer than this will be removed.

Returns:

nothing, modifies spiketrains in place.

SpikeTrainUtilities.filter_out_small_isi! — Method

function filteroutsmallisi!(spiketrains::Vector{<:AbstractVector{<:Real}}, dtmin::Real)

Filter out spikes (events) that are closer than dt_min from each other in the given spiketrains. This modifies the input spiketrains in place, removing spikes that are too close together.

Arguments:

spiketrains::Vector{<:AbstractVector{<:Real}}: A vector of spike trains, where each train is an array of spike times.

dt_min::Real: The minimum time difference between spikes. Spikes closer than this will be removed.

Returns:

nothing, modifies spiketrains in place.

SpikeTrainUtilities.get_instantaneous_firing_rates — Method

function getinstantaneousfiringrates(spiketrains::SpikeTrains; dtmin::Real=0.0)

Calculate the instantaneous firing rates, iFR, defined as 1/ISI, for each spike train in the spiketrains object. The iFR time is defined as the time of the right spike in the ISI.

Arguments:

spiketrains::SpikeTrains: An instance of SpikeTrains, which contains multiple spike trains.

dt_min::Real: The minimum time difference between spikes. Spikes closer than this will be removed before calculating iFR, default is 0.0, which means no filtering.

Returns:

iFRTrains::SpikeQuantity: An instance of SpikeQuantity containing the instantaneous firing rates for each spike train. The event_times are the times of the left spikes in the ISI, and ys are the corresponding iFR values.

SpikeTrainUtilities.get_line_segments — Method

function getlinesegments(spiketrains::SpikeTrains, tstart::Real,tend::Real; heightscaling::Real=0.7, neurons::Union{Vector{Int},Nothing}=nothing, offset::Real = 0.0, maxspikes::Real=1E6)

Returns point coordinates that can be used for line segments for a raster plot of the spiketrain

Arguments

spiketrains::SpikeTrains : the spiketrains

t_start::Real : the starting time of the raster

t_end::Real : the ending time of the raster

height_scaling::Real=0.7 : how tall each spike is, 1.0 is the full height

neurons::Union{Vector{Int},Nothing}=nothing : the neurons to plot, if nothing all neurons are plotted

neuron_offset::Real=0.0 : the y-offset of the raster

time_offset::Real : the time offset of the raster, default is t_start

max_spikes::Real=1E6 : the maximum number of spikes that can be plotted.

Returns

points_linesegment : a vector of tuples of two floats, (t1,y1a),(t1,y1b),(t2,y2a),(t2,y2b),... where t1,t2 are the times of the spikes and y1a,y1b,y2a,y2b are the y-coordinates of the line segments, representing neurons on the y axis. This can be used directly as argument to linesegments! in Makie.

SpikeTrainUtilities.line_segments_to_centers_and_size — Method

function linesegmentstocentersand_size(points::Vector{Tuple{R,R}}) where R<:Real

Calculates the centers and sizes of the segments obtained from get_line_segments.

Arguments

points::Vector{Tuple{R,R}} : the points obtained from get_line_segments

Returns

x_centers::Vector{R} : the x-coordinates of the centers of the segments

y_centers::Vector{R} : the y-coordinates of the centers of the segments

size::R : the size of the segments in y direction

SpikeTrainUtilities.line_segments_to_xynans — Method

function linesegmentstoxynans(points::Vector{Tuple{R,R}}) where R<:Real Converts segments obtained from `getline_segments` to x and y vectors, using NaN to disconnect segments.

Arguments

points::Vector{Tuple{R,R}} : the points obtained from get_line_segments

Returns

x_plot_all::Vector{R} : x-coordinates of the points

y_plot_all::Vector{R} : y-coordinates of the points, with NaN to disconnect segments

SpikeTrainUtilities.numerical_rates — Method

numericalrates(S::SpikeTrains{R,N};tstart::R=0.0,t_end::R=Inf) where {R,N}

Compute the numerical rates of the spike trains in S over the interval [t_start,t_end] if specified. If not, uses the interval [S.t_start,S.t_end].

SpikeTrainUtilities.sort_units! — Method

function sortunits!(spk::SpikeTrains{R,N,T},sortingdict::Dict{T,I};strict_dictionary::Bool=true) where {R,N,T,I<:Integer}

Modifies the spk object in place by sorting its units according to the provided sorting_dict. The sorting_dict maps original unit identifiers (of type T) to ranking numbers. The dictionary must contain all keys that correspond to the units in spk. If strict_dictionary is true, then the dictionary must contain exactly the units present in spk and vice versa. If strict_dictionary is false, then the dictionary can contain extra keys.

Arguments:

spk::SpikeTrains{R,N,T}: The spike trains object to be sorted.

sorting_dict::Dict{T,RA}: A dictionary mapping unit identifiers to a ranking number.

strict_dictionary::Bool: If true, the function will check that the sorting_dict contains exactly the units present in spk and vice versa. If false, it will only check that the keys in sorting_dict are a subset of the units in spk, and use the values as rankings for sorting.

Returns:

nothing, modifies spk in place.