SpikeTrainUtilities

Documentation for SpikeTrainUtilities.

SpikeTrainUtilities.SpikeTrains

SpikeTrainUtilities.bin_spikes

SpikeTrainUtilities.draw_spike_raster

SpikeTrainUtilities.numerical_rates

SpikeTrainUtilities.SpikeTrains — Type

SpikeTrains(nunits::N,trains::Vector{Vector{R}},tstart::R,t_end::R) where {R,N}

Constructs a SpikeTrains object.

SpikeTrainUtilities.bin_spikes — Method

bin_spikes(Y::Vector{R},dt::R,Tend::R;Tstart::R=0.0) where R

Arguments

Y::Vector{<:Real} : vector of spike times

dt::Real : time bin size

Tend::Real : end time for the raster

Optional argument

Tstart::Real=0.0 : start time for the raster

Returns

timesbin::Vector{R} : `timesbin[k]is the midpoint of the timebink(i.e.Tstart + (k-1/2)*dt`)

binned_spikes::Vector{<:Integer} : binned_spikes[k] is the number of spikes that occur in the timebin k (i.e. between Tstart + (k-1)*dt and Tstart + k*dt)

SpikeTrainUtilities.draw_spike_raster — Method

drawspikeraster(trains::Vector{Vector{Float64}}, dt::Real,Tend::Real; Tstart::Real=0.0, spikesize::Integer = 5, spikeseparator::Integer = 1, backgroundcolor::Color=RGB(1.,1.,1.), spikecolors::Union{C,Vector{C}}=RGB(0.,0.0,0.0), max_size::Real=1E4) where C<:Color

Draws a matrix that contains the raster plot of the spike train.

Arguments

Trains : Vector of spike trains. The order of the vector corresponds to the order of the plot. First element is at the top, second is second row, etc.

dt : time interval representing one horizontal pixel

Tend : final time to be considered

Optional arguments

Tstart::Real : starting time

spike_size::Integer : heigh of spike (in pixels)

spike_separator::Integer : space between spikes, and vertical padding

background_color::Color : self-explanatory

spike_colors::Union{Color,Vector{Color}} : if a single color, color of all spikes, if vector of colors, color for each neuron (length should be same as number of neurons)

max_size::Integer : throws an error if image is larger than this number (in pixels)

Returns

raster_matrix::Matrix{Color} you can save it as a png file

SpikeTrainUtilities.numerical_rates — Method

numericalrates(S::SpikeTrains{R,N};tstart::R=0.0,t_end::R=Inf) where {R,N}

Compute the numerical rates of the spike trains in S over the interval [t_start,t_end] if specified. If not, uses the interval [S.t_start,S.t_end].